more encapsulation
